🔊 Running `phi` and Configuring logging behavior
`phi` may now be ran as a script, performing the preprocessing steps for
joining on common columns.

Added argument parsing, logging, and configuration file options to the
`phi.__main__` so that options for various settings can be used.

Updated `phi/unittests/config_tests/test_initialize_settings.py` since the
`target: "..."` option in `get_config.py` was changed to correctly add the
path to the beginning.